pyproject.toml
[project]
name = "codesnippet-hub"
version = "0.1.0"
description = "A personal code snippet manager with syntax highlighting"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.8.1"
dependencies = [
    "flet>=0.24.1",
    "pyperclip>=1.8.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]